+++ README.md
@@ -51,7 +51,6 @@
 # TODO List
 
 - Реализация бизнес-логики на основании chain-of-responsibility
-- Интеграция с Postgre (Jooq )
 - Реализация gradle Tasks для самостоятельной генерации описания БД
 - Интеграция с TestContainers
 - Упаковка решения в контейнеры
