+++ project-build/postgre-repo/src/test/kotlin/ImageRepoSeverDBTest.kt
@@ -6,6 +6,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.test.TestScope
 import kotlinx.coroutines.test.runTest
 import kotlinx.coroutines.withContext
+import liquibase.Liquibase
 import org.junit.Test
 import repo.SQLParams
 import ru.otus.kotlin.course.common.models.PsImage
@@ -19,6 +20,8 @@ import kotlin.test.assertIs
 import kotlin.time.Duration.Companion.minutes
 
 
+
+
 class ImageRepoSeverDBTest() {
 
     private fun initRepo (repo: ImageRepoDB, objects: List<PsImage>)
@@ -30,102 +33,17 @@ class ImageRepoSeverDBTest() {
         }
     }
 
-    val params = SQLParams(host = "localhost")
+
 
     @Test
-    fun readImageTest () {
-        val repo = ImageRepoDB( params = params,  randomId = { getDefaultId() } )
-        initRepo(repo, listOf(PsImageStubsItems.FULL_TO_PSIMAGE))
-        runRepoTest {
-            val res = repo.readImage(DBImageId(getDefaultId()), false)
-            assertIs<DBGetImage>(res)
-            val obj = res.image
-            assertEquals<PsImage>(PsImageStubsItems.FULL_TO_PSIMAGE, obj)
-            // abnormal case test in deleteImageTest
-        }
-    }
+    fun executeTests () {
+        val params = SQLParams(host = "localhost")
+        val runner: TestRunner = TestRunner(params)
 
-    @Test
-    fun updateImageTest () {
-        val repo = ImageRepoDB( params = params,  randomId = { getDefaultId() } )
-        initRepo(repo, listOf(PsImageStubsItems.FULL_TO_PSIMAGE))
-        runRepoTest {
-            val image = PsImageStubsItems.FULL_TO_PSIMAGE.copy(imageUrl = "www.kremlin.ru")
-            repo.updateImage(DBImageRequest( image),)
-            val res = repo.readImage(DBImageId(getDefaultId()), false)
-            assertIs<DBGetImage>(res)
-            val obj = res.image
-            assertEquals<PsImage>(image, obj)
-
-            // wrong ID
-            val WRONG_ID = "321"
-            val wrongImage = PsImageStubsItems.FULL_TO_PSIMAGE.copy(id = PsImageId(WRONG_ID))
-            val notRes = repo.updateImage(DBImageRequest( wrongImage))
-            assertIs<DBError>(notRes)
-            assertEquals(errorNotFound(WRONG_ID), notRes)
-        }
-    }
-
-    @Test
-    fun createImageTest () {
-        val repo = ImageRepoDB( params = params,  randomId = { getDefaultId() } )
-        initRepo(repo, emptyList())
-        runRepoTest {
-            val img = PsImageStubsItems.FULL_TO_PSIMAGE.copy()
-            val BYTES = byteArrayOf (0x30, 0x31, 0x32)
-            img.file = BYTES
-            val cr = repo.createImage(DBImageRequest(img))
-            assertIs<DBGetImage>(cr)
-            val res = repo.readImage(DBImageId(cr.image.id.asString()), true)
-            assertIs<DBGetImage>(res)
-            val obj = res.image
-            val image = PsImageStubsItems.FULL_TO_PSIMAGE.copy(id = cr.image.id )
-            assertEquals<PsImage>(image, obj)
-            assertTrue(obj.file.contentEquals(BYTES))
-        }
-    }
-
-    @Test
-    fun deleteImageTest () {
-        val repo = ImageRepoDB( params = params,  randomId = { getDefaultId() } )
-        initRepo(repo, listOf(PsImageStubsItems.FULL_TO_PSIMAGE))
-        runRepoTest {
-            val res = repo.deleteImage(DBImageId(getDefaultId()))
-            assertIs<DBGetImage>(res)
-            val obj = res.image
-            assertEquals<PsImage>(PsImageStubsItems.FULL_TO_PSIMAGE, obj)
-            // try delete again - faile
-            val deleteRes = repo.deleteImage(DBImageId(getDefaultId()))
-            assertIs<DBError>(deleteRes)
-            assertEquals(errorNotFound(getDefaultId()), deleteRes)
-            // try read - fail
-            val readRes = repo.readImage(DBImageId(getDefaultId()), false)
-            assertIs<DBError>(readRes)
-            assertEquals(errorNotFound(getDefaultId()), readRes)
-        }
-    }
-
-    @Test
-    fun searchImagesTest() {
-        val repo = ImageRepoDB(params = params, randomId = { getDefaultId() })
-        val image1 = PsImageStubsItems.FULL_TO_PSIMAGE.copy(title = "CriteriaA", id = PsImageId("1234"))
-        val image2 = PsImageStubsItems.FULL_TO_PSIMAGE.copy(title = "CriteriaB")
-        initRepo(repo, listOf(image1, image2 ))
-        runRepoTest {
-            val resTwo = repo.searchImages(DBImageSearchFilter("Criteria"))
-            assertIs<DBGetImages>(resTwo)
-            assertTrue(resTwo.images.size == 2)
-            assertEquals<PsImage>(image1, resTwo.images[0])
-            assertEquals<PsImage>(image2, resTwo.images[1])
-
-            val resOne = repo.searchImages(DBImageSearchFilter("CriteriaA"))
-            assertIs<DBGetImages>(resOne)
-            assertTrue(resOne.images.size == 1)
-            assertEquals<PsImage>(image1, resOne.images[0])
-
-            val resNone = repo.searchImages(DBImageSearchFilter("NOT_FOUND"))
-            assertIs<DBGetImages>(resNone)
-            assertTrue(resNone.images.size == 0)
-        }
+        runner.createImageTest()
+        runner.readImageTest()
+        runner.updateImageTest()
+        runner.deleteImageTest()
+        runner.searchImagesTest()
     }
 }
